changed amount to double and fixed delete shoppinglist items so it takes a list

+++ src/main/java/edu/ntnu/idatt2106_09/backend/model/Fridge.java
@@ -43,7 +43,7 @@ public class Fridge {
             fetch = FetchType.EAGER)
     private Set<GroceryItemFridge> groceries = new HashSet<>();
 
-    public void addGroceryItem(GroceryItem groceryItem, int amount) {
+    public void addGroceryItem(GroceryItem groceryItem, double amount) {
         GroceryItemFridge groceryItemFridge = new GroceryItemFridge(this, groceryItem, amount);
         groceries.add(groceryItemFridge);
     }
@@ -63,7 +63,7 @@ public class Fridge {
         }
     }
 
-    public void updateGroceryItem(GroceryItem groceryItem, int amount, int actualShelfLife, boolean negative, LocalDateTime timestamp) {
+    public void updateGroceryItem(GroceryItem groceryItem, double amount, int actualShelfLife, boolean negative, LocalDateTime timestamp) {
         for (GroceryItemFridge groceryItemFridge : groceries) {
             if (groceryItemFridge.getFridge().equals(this) &&
                     groceryItemFridge.getGroceryItem().equals(groceryItem) &&
